feat: improve Quarkus component detection
Quarkus projects often have lots of modules that are not intended to be
components, but still have io.quarkus dependencies.

This commit will only mark a module as a Quarkus component if it
explicitly references the Quarkus plugin that creates a runnable
application.

In addition if a Java project contains components that explicitly
specify frameworks then modules without frameworks will be removed.

This will help with the usability issues identified in RHTAPBUGS-552